#9 [RENAME} 공통 타입 정의 파일명 변경
