+++ src/components/ListView/IssueItem.tsx
@@ -1,4 +1,4 @@
-import type { IssueItemProps } from '../../types/goal';
+import type { IssueItemProps } from '../../types/listItem';
 import dateIcon from '../../assets/icons/date.svg';
 import grayIcon from '../../assets/icons/gray.svg';
 import goalIcon from '../../assets/icons/goal.svg';
